Renamed files and moved functions
